==> bert/00_crawl_researvedwords.ipynb <==
# pip install beautifulsoup4

import requests
from bs4 import BeautifulSoup
import pickle

url = 'https://github.com/AnanthaRajuC/Reserved-Key-Words-list-of-various-programming-languages/blob/master/language-files/Python%20Reserved%20Words.md'

response = requests.get(url)

if response.status_code == 200:
    html = response.text
    soup = BeautifulSoup(html, 'html.parser')
    title = soup.select_one('#repo-content-pjax-container > react-app > div > div > div.Box-sc-g0xbh4-0.fSWWem > div > div > div.Box-sc-g0xbh4-0.emFMJu > div.Box-sc-g0xbh4-0.hlUAHL > div > div:nth-child(3) > div.Box-sc-g0xbh4-0.iJmJly > div > div.Box-sc-g0xbh4-0.ytOJl > section > div > article')
    print(title)
else : 
    print(response.status_code)

reserved_list = ['and',
'assert',
'break',
'class',
'continue',
'def',
'del',
'elif',
'else',
'except',
'exec',
'finally',
'for',
'from',
'global',
'if',
'import',
'in',
'is',
'lambda',
'not',
'or',
'pass',
'print',
'raise',
'return',
'try',
'while',
'Float',
'Int',
'Numeric',
'Oxphys',
'array',
'close',
'float',
'int',
'input',
'open',
'range',
'type',
'write',
'zeros',
'acos',
'asin',
'atan',
'cos',
'exp',
'fabs',
'floor',
'log',
'log10',
'pi',
'sin',
'sqrt',
'tan'
'False',               'break',               'for',                 'not',
'None',                'class',               'from',                'or',
'True',                'continue',            'global',              'pass',
'__peg_parser__',      'def',                 'if',                  'raise',
'and',                 'del',                 'import',              'return',
'as',                  'elif',                'in',                  'try',
'assert',              'else',                'is',                  'while',
'async',               'except',              'lambda',              'with',
'await',               'finally',             'nonlocal',            'yield',
'quot' ,                'lt',                 'gt'

]

# 전체 데이터셋 
with open('../../data/reserved_list.pkl', 'wb') as f:
    pickle.dump(reserved_list, f)



help("keywords")
