==> bert/1_test.ipynb <==
https://zerojsh00.github.io/posts/BERTopic/
https://towardsdatascience.com/topic-modeling-with-bert-779f7db187e6
https://arxiv.org/pdf/2203.05794.pdf

# ! which python

# pip install bertopic

# pip uninstall umap


# pip install hdbscan

# pip install umap-learn

import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import numpy as np
import pickle
import re
import umap.umap_ as umap
import hdbscan

# 파일 불러오기
with open('../../data/bert_df.pkl', 'rb') as f:
    df = pickle.load(f)


df.info()

# extract the  columns for text analysis
## q_body : question body
## a_body : answer body
df_qna = df[['q_id','a_id','q_text', 'a_text']]

df_qna.info()

def cleanhtml(raw_html):
  # 1.Source code in python language is hard to understand, so replace all the <code> tag first
  cleantext_1 = re.findall(r'(?<=\<code>)(.*?)(?=<\/code>)', raw_html.replace('\n', '_**_'))
  cleantext_1 = [x.replace('_**_', '\n') for x in cleantext_1]
  # 2. replace html tags
  # <p>
  tag_re = re.compile('<.*?>')
  cleantext_2 = [re.sub(tag_re, '', x) for x in cleantext_1]
  return cleantext_2

# apply the function, cleanhtml to the question and body text
df_qna.loc[:, 'q_prep_text'] = df_qna['q_text'].apply(cleanhtml)
df_qna.loc[:, 'a_prep_text'] = df_qna['a_text'].apply(cleanhtml)

df_qna.loc[:, [ 'q_id','q_prep_text', 'a_prep_text']] 

df_q_src = df_qna[['q_id', 'q_prep_text']].apply(pd.Series.explode)
df_a_src = df_qna[['a_id', 'a_prep_text']].apply(pd.Series.explode)

df_q_src = df_q_src.reset_index(drop=True)
df_a_src = df_a_src.reset_index(drop=True)
df_q_src.dropna(inplace=True)
df_a_src.dropna(inplace=True)

# df_q_src[df_q_src['q_prep_text']==''] 공백있음
df_q_src = df_q_src[~(df_q_src['q_prep_text']=='')]

df_q_src

df_q_src['q_prep_text'] = df_q_src['q_prep_text'].str.replace('\n', ' ')
df_a_src['a_prep_text'] = df_a_src['a_prep_text'].str.replace('\n', ' ')

print(df_q_src['q_prep_text'].isna().sum())
print(df_a_src['a_prep_text'].isna().sum())

src_list = df_q_src.loc[:, 'q_prep_text'].values

src_list[0]

from transformers import AutoTokenizer, AutoModel
import torch

tokenizer = AutoTokenizer.from_pretrained("microsoft/codebert-base")
model = AutoModel.from_pretrained("microsoft/codebert-base")

code_tokens=tokenizer.tokenize(src_list[0:10])
# src_list[0:10] 한번에 대량으로 넣으면 안됨....

tokens_ids=tokenizer.convert_tokens_to_ids(code_tokens)

context_embeddings=model(torch.tensor(tokens_ids)[None,:])[0]


print(tokenizer.cls_token)
print(tokenizer.sep_token)
print(tokenizer.eos_token)

# context_embeddings = context_embeddings.reshape(-1,768)

context_embeddings.shape


umap_embeddings = umap.UMAP(n_neighbors=15,
                            n_components=5,
                            metric='cosine').fit_transform(context_embeddings.detach().numpy())




cluster = hdbscan.HDBSCAN(min_cluster_size=15,
                          metric='euclidean',
                          cluster_selection_method='eom').fit(umap_embeddings)

cluster.labels_

from openai import OpenAI
client = OpenAI(api_key = 'sk-proj-Mm0FIOmjCUgXc2T40TarT3BlbkFJ7YO5SE9kJokyhFrRxEiB')

completion = client.chat.completions.create(
  model="gpt-3.5-turbo",
  messages=[
    {"role": "system", "content": "You are a poetic assistant, skilled in explaining complex programming concepts with creative flair."},
    {"role": "user", "content": """I have a topic that contains the following documents:
[DOCUMENTS]
The topic is described by the following keywords: [KEYWORDS]

Based on the information above, extract a short but highly descriptive topic label of at most 5 words. Make sure it is in the following format:
topic: <topic label>"""}
  ]
)

print(completion.choices[0].message)

len(["""I have a topic that contains the following documents:
[DOCUMENTS]
The topic is described by the following keywords: [KEYWORDS]

Based on the information above, extract a short but highly descriptive topic label of at most 5 words. Make sure it is in the following format:
topic: <topic label>"""])
