==> bert/2_test.ipynb <==
https://zerojsh00.github.io/posts/BERTopic/
https://towardsdatascience.com/topic-modeling-with-bert-779f7db187e6
https://arxiv.org/pdf/2203.05794.pdf

https://dacon.io/competitions/official/235900/codeshare?page=1&dtype=recent&ptype=pub&keyword=

import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import numpy as np
import pickle
import re
import umap.umap_ as umap
import torch
from transformers import RobertaTokenizer, RobertaConfig, RobertaModel
from collections.abc import Iterable

# 파일 불러오기
with open('../../data/bert_df.pkl', 'rb') as f:
    df = pickle.load(f)


df.info()

# extract the  columns for text analysis
## q_body : question body
## a_body : answer body
df_qna = df[['q_id','a_id','q_text', 'a_text']]

df_qna.info()

def cleanhtml(raw_html):
  # 1.Source code in python language is hard to understand, so replace all the <code> tag first
  cleantext_1 = re.findall(r'(?<=\<code>)(.*?)(?=<\/code>)', raw_html.replace('\n', '_**_'))
  cleantext_1 = [x.replace('_**_', '\n') for x in cleantext_1]
  # 2. replace html tags
  # <p>
  tag_re = re.compile('<.*?>')
  cleantext_2 = [re.sub(tag_re, '', x) for x in cleantext_1]
  return cleantext_2

# apply the function, cleanhtml to the question and body text
df_qna.loc[:, 'q_prep_text'] = df_qna['q_text'].apply(cleanhtml)
df_qna.loc[:, 'a_prep_text'] = df_qna['a_text'].apply(cleanhtml)

df_qna.loc[:, [ 'q_id','q_prep_text', 'a_prep_text']] 

df_q_src = df_qna[['q_id', 'q_prep_text']].apply(pd.Series.explode)
df_a_src = df_qna[['a_id', 'a_prep_text']].apply(pd.Series.explode)

df_q_src = df_q_src.reset_index(drop=True)
df_a_src = df_a_src.reset_index(drop=True)

df_q_src.dropna(inplace=True)
df_q_src['q_prep_text_non'] = df_q_src['q_prep_text'].str.replace('\n', ' ')

df_q_src['q_prep_text_non']

src = df_q_src['q_prep_text_non'].tolist()
# df['sentiments'].values.tolist()


print(type(src))
print(type(src[0]))
print(len(src))

src = src[:10]

Pre-calculate Embeddings

from transformers import AutoTokenizer, AutoModel
import torch
tokenizer = AutoTokenizer.from_pretrained("microsoft/codebert-base")
embedding_model = AutoModel.from_pretrained("microsoft/codebert-base")

# https://stackoverflow.com/questions/63517293/valueerror-textencodeinput-must-be-uniontextinputsequence-tupleinputsequence/63870843#63870843
code_tokens=tokenizer(src, truncation=True, padding=True)
tokens_ids=tokenizer.convert_tokens_to_ids(code_tokens)
embeddings=embedding_model(torch.tensor(tokens_ids)[None,:])[0]
# embeddings = embeddings.reshape(-1,768)

# tokenizer = AutoTokenizer.from_pretrained("microsoft/codebert-base")
# model = AutoModel.from_pretrained("microsoft/codebert-base")
# code_tokens=tokenizer.tokenize(src[0])

embeddings.shape

Train

from bertopic import BERTopic

topic_model = BERTopic(

  # Pipeline models
  embedding_model=embedding_model,
  # umap_model=umap_model,
  # hdbscan_model=hdbscan_model,
  # vectorizer_model=vectorizer_model,
  # representation_model=representation_model,

  # Hyperparameters
  top_n_words=10,
  verbose=True
)



type(embeddings)

# Convert the tensor to a NumPy array
embeddings_numpy = embeddings.detach().numpy()
# embeddings_numpy = embeddings_numpy[0]
print(type(embeddings_numpy))

test = 'The problem of statistical learning is to construct a predictor of a random\nvariable $Y$ as a function of a related random variable $X$ on the basis of an\ni.i.d. training sample from the joint distribution of $(X,Y)$. Allowable\npredictors are drawn from some specified class, and the goal is to approach\nasymptotically the performance (expected loss) of the best predictor in the\nclass. We consider the setting in which one has perfect observation of the\n$X$-part of the sample, while the $Y$-part has to be communicated at some\nfinite bit rate. The encoding of the $Y$-values is allowed to depend on the\n$X$-values. Under suitable regularity conditions on the admissible predictors,\nthe underlying family of probability distributions and the loss function, we\ngive an information-theoretic characterization of achievable predictor\nperformance in terms of conditional distortion-rate functions. The ideas are\nillustrated on the example of nonparametric regression in Gaussian noise.\n'

test

topics, probs = topic_model.fit_transform(src[0:10], embeddings_numpy)

type(embeddings_numpy)

if isinstance(src[0:2], pd.DataFrame):
        raise TypeError("Make sure to supply a list of strings, not a dataframe.")
elif isinstance(src[0:2], Iterable) and not isinstance(src[0:2], str):
    if not any([isinstance(doc, str) for doc in src[0:2]]):
        raise TypeError("Make sure that the iterable only contains strings.")
else:
    raise TypeError("Make sure that the documents variable is an iterable containing strings only.")


test

if isinstance(test, pd.DataFrame):
        raise TypeError("Make sure to supply a list of strings, not a dataframe.")
elif isinstance(test, Iterable) and not isinstance(test, str):
    if not any([isinstance(doc, str) for doc in test]):
        raise TypeError("Make sure that the iterable only contains strings.")
else:
    raise TypeError("Make sure that the documents variable is an iterable containing strings only.")



if not isinstance(src[0], str):
    print("aaa")
else :
    print("bbb")

isinstance(src[0], str)


embeddings_numpy.shape

df_q_src

df_q_src= df_q_src['q_prep_text_non'].apply(tokenizer.tokenize)
