==> bert/13_baf_gpt.ipynb <==
2021 ~ 2023년  
김태균 교수님 피드백대도 다달이 10%의 질문을 샘플링 해서 해보자

import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import numpy as np
import pickle
import re
from umap import UMAP
from hdbscan import HDBSCAN
import torch
from transformers import RobertaTokenizer, RobertaConfig, RobertaModel
from collections.abc import Iterable
from transformers import AutoTokenizer, AutoModel
from bertopic import BERTopic
import torch
from collections import deque
from bertopic.representation import KeyBERTInspired
from tqdm import tqdm
from sklearn.cluster import KMeans
import calendar
from bertopic.vectorizers import ClassTfidfTransformer
from sklearn.feature_extraction.text import CountVectorizer
import openai
from bertopic.representation import KeyBERTInspired, MaximalMarginalRelevance, OpenAI, PartOfSpeech
from itertools import product

# sample_df
# 전체 데이터셋 
with open('../../data/sample_df.pkl', 'rb') as f:
    sample_df = pickle.load(f)

# 파일 불러오기
with open('../../data/bert_src_df.pkl', 'rb') as f:
    bert_src_df = pickle.load(f)


bert_src_df.info()
# 74308951

bert_src_df[bert_src_df['q_id'] ==74308951]

cond1 = bert_src_df['q_creationdate']<='2022-11-30'
cond2 = bert_src_df['q_creationdate']>='2021-11-30'
cond3 = bert_src_df['a_id'].isna()
cond4 = bert_src_df['tags'].str.contains('python')
cond5 = bert_src_df['q_creationdate']>='2022-12-01'

bert_src = bert_src_df.loc[cond2 & cond4, :]

bert_src.head()

def cleanhtml(raw_html):
  # 1.Source code in python language is hard to understand, so replace all the <code> tag first
  cleantext_1 = re.findall(r'(?<=\<code>)(.*?)(?=<\/code>)', raw_html.replace('\n', '_**_'))
  cleantext_1 = [x.replace('_**_', '\n') for x in cleantext_1]
  # 2. replace html tags
  # <p>
  tag_re = re.compile('<.*?>')
  cleantext_2 = [re.sub(tag_re, '', x) for x in cleantext_1]
  return cleantext_2

# apply the function, cleanhtml to the question and body text
bert_src.loc[:, 'q_prep_text'] = bert_src['body'].apply(cleanhtml)

bert_src = bert_src.reset_index(drop=True)

bert_src_ex = bert_src[['q_id', 'q_prep_text']].apply(pd.Series.explode)

bert_src_ex.dropna(inplace=True)

bert_src = pd.merge(bert_src_ex, bert_src[['q_id', 'q_creationdate']], how ='left', left_on = 'q_id', right_on = 'q_id')

bert_src['q_year'] = pd.to_datetime(bert_src['q_creationdate']).dt.year
bert_src['q_month'] = pd.to_datetime(bert_src['q_creationdate']).dt.month
bert_src['q_day'] = pd.to_datetime(bert_src['q_creationdate']).dt.day

# PREPROCESSING FOR CODE SCRIPT
def preprocess_script(script):
    new_script = deque()
    old_script = script.split('\n')
    for line in old_script:
        if line.lstrip().startswith('#'): # 주석으로 시작되는 행 skip
            continue
        line = line.rstrip()
        if '#' in line:
            line = line[:line.index('#')] # 주석 전까지 코드만 저장
        line = line.replace('\n','') # 개행 문자를 모두 삭제함
        line = line.replace('    ','\t') # 공백 4칸을 tab으로 변환
        
        if line == '': # 전처리 후 빈 라인은 skip
            continue
        
        new_script.append(line)

        
    new_script = '\n'.join(new_script) # 개행 문자로 합침
    #이렇게 할지 ..... 고민을 해봐야 할것같음... 전처리......
    new_script = re.sub('("""[\w\W]*?""")', '', new_script)
    new_script = re.sub("('''[\w\W]*?''')", '', new_script)
    new_script = re.sub('/^(http?|https?):\/\/([a-z0-9-]+\.)+[a-z0-9]{2,4}.*$/', '', new_script)
    
    return new_script


bert_src['q_prep_text_non'] = bert_src['q_prep_text'].apply(preprocess_script)

# bert_src.loc[bert_src['q_prep_text_non']!= bert_src['q_prep_text'], ['q_prep_text', 'q_prep_text_non']].head(100).to_csv('chk_preprocess.csv')

bert_src.iloc[51:54, :]

# CodeBERT 모델과 토크나이저 로드
model_name = "microsoft/codebert-base"
tokenizer = AutoTokenizer.from_pretrained(model_name)
model = AutoModel.from_pretrained(model_name)

# 임베딩 함수 정의
def embed_text(text):
    # 텍스트를 토크나이저를 사용해 인코딩
    encoded_input = tokenizer(text, padding=True, truncation=True, return_tensors='pt')
    
    # 모델 추론 시 기울기를 계산하지 않도록 설정 > 기울기를 계산하여 파라미터를 업데이트 하지 않는다고?
    with torch.no_grad():
        # 인코딩된 입력을 모델에 전달하여 출력 생성(key:value로 해서 model output에 전달)
        model_output = model(**encoded_input)
    
    # 모델의 출력에서 [CLS] 토큰의 임베딩을 반환
    # 입력시퀀스의 각 토큰의 마지막 은닉상태를 가져오기 위해 last_hidden_state에 접근
    # [batch_size, sequence_length, hidden_size]의 텐서
    # 중간 파라미터를 0으로 두어 첫번째 토큰인 cls를 가져옴
    # 그래서 결과는 [batch_size, hidden_size가 됨]
    return model_output.last_hidden_state[:, 0, :]

years = [2021,2022, 2023]
months = [x for x in range(1, 13)]
ym = list(product(years, months))

ym[10:-1]

bert_src

q_id_df  = pd.DataFrame()

for year, month in ym[10:-1]:
    startend = calendar.monthrange(year, month)
    

    
    cond1 = (bert_src['q_year'] ==year)
    cond2 = (bert_src['q_month'] ==month)

    q_id_list = bert_src.loc[cond1&cond2, 'q_id'].unique()
    num_of_sample = int(len(q_id_list)*0.1)
    print(year, month, len(q_id_list))
    q_id_df = pd.concat([q_id_df, pd.DataFrame(q_id_list).sample(num_of_sample,replace=False)], axis = 0)



q_id_df.columns  = ['q_id']

q_id_df['q_id'].values

embed_text_list = []
for year, month in ym[10:-1]:
    startend = calendar.monthrange(year, month)
    day = startend[1]

    print(year, month, day) 
    cond1 = (bert_src['q_year'] ==year)
    cond2 = (bert_src['q_month'] ==month)
    # cond3 = (bert_src['q_day'] ==day)
    cond4 = (bert_src['q_id'].isin(q_id_df['q_id'].values))
    src = bert_src.loc[cond1&cond2&cond4, 'q_prep_text_non'].tolist()
    for i in tqdm(range(len(src))) : 
        embed_text_list.append(embed_text(src[i]))

data = pd.DataFrame()
for year, month in ym[10:-1]:
    startend = calendar.monthrange(year, month)
    day = startend[1]

    print(year, month, day) 
    cond1 = (bert_src['q_year'] ==year)
    cond2 = (bert_src['q_month'] ==month)
    # cond3 = (bert_src['q_day'] ==day)
    cond4 = (bert_src['q_id'].isin(q_id_df['q_id'].values))
    data = pd.concat([data, bert_src.loc[cond1&cond2&cond4, :]], axis = 0)


# 파일 불러오기
with open('../../data/reserved_list.pkl', 'rb') as f:
    reserved_list = pickle.load(f)


src = data['q_prep_text'].to_list()

# CodeBERT 모델과 토크나이저 로드
model_name = "microsoft/codebert-base"
tokenizer = AutoTokenizer.from_pretrained(model_name)
model = AutoModel.from_pretrained(model_name)

# 모든 텍스트에 대한 임베딩 계산
embeddings = torch.cat(embed_text_list, dim=0).numpy()  # 텐서로 연결 후 numpy 배열로 변환

embeddings.shape

umap_model = UMAP(n_neighbors=10, n_components=5, min_dist=0.0, metric='cosine', random_state=42)

hdbscan_model = HDBSCAN(min_cluster_size=20, metric='euclidean', cluster_selection_method='eom', prediction_data=True)
cluster_model = KMeans(n_clusters=10)

# KeyBERT
keybert_model = KeyBERTInspired()

# Part-of-Speech
# pos_model = PartOfSpeech("en_core_web_sm")

# MMR
mmr_model = MaximalMarginalRelevance(diversity=0.3)

# GPT-3.5
prompt = """
I have a topic that contains the documents: [REPRESENTATIVE_DOCUMENT]
The topic is described by the following keywords: [KEYWORDS]

Based on the information above, extract a short but highly descriptive topic label of at most 5 words. Make sure it is in the following format:
topic: <topic label>
"""
client = openai.OpenAI(api_key="sk-proj-Mm0FIOmjCUgXc2T40TarT3BlbkFJ7YO5SE9kJokyhFrRxEiB")
openai_model = OpenAI(client, model="gpt-3.5-turbo", exponential_backoff=True, chat=True, prompt=prompt)

# All representation models
representation_model = {
    "KeyBERT": keybert_model,
    "OpenAI": openai_model,  # Uncomment if you will use OpenAI
    "MMR": mmr_model
    # ,
    # "POS": pos_model
}

vectorizer_model = CountVectorizer(stop_words=reserved_list, min_df=2, ngram_range=(1, 2))

ctfidf_model = ClassTfidfTransformer(reduce_frequent_words=True)


# BERTopic 모델 초기화 및 훈련
topic_model = BERTopic( embedding_model=model,
                        umap_model=umap_model,
                        # hdbscan_model=hdbscan_model,
                        hdbscan_model=cluster_model,
                        ctfidf_model=ctfidf_model,
                        vectorizer_model = vectorizer_model,
                        representation_model=representation_model)  # 임베딩 모델 사용을 비활성화
topics, probabilities = topic_model.fit_transform(src, embeddings)


# 결과 출력
print(topic_model.get_topic_info())  # 토픽 정보 출력

topic_model.get_topic_info()

topic_model.get_topic(0)

topic_model.get_topic(1)

topic_model.get_topic(6, full=True)

# # Label the topics yourself
# topic_model.set_topic_labels({1: "Space Travel", 7: "Religion"})

# # or use one of the other topic representations, like KeyBERTInspired
# keybert_topic_labels = {topic: " | ".join(list(zip(*values))[0][:3]) for topic, values in topic_model.topic_aspects_["KeyBERT"].items()}
# topic_model.set_topic_labels(keybert_topic_labels)

# # or ChatGPT's labels
# chatgpt_topic_labels = {topic: " | ".join(list(zip(*values))[0]) for topic, values in topic_model.topic_aspects_["OpenAI"].items()}
# chatgpt_topic_labels[-1] = "Outlier Topic"
# topic_model.set_topic_labels(chatgpt_topic_labels)

# topic_distr, _ = topic_model.approximate_distribution(data, window=8, stride=4)

# pip install nbformat>=4.2.0

# Visualize the topic-document distribution for a single document
topic_model.visualize_topics()

topic_model.visualize_hierarchy()

reduced_embeddings = UMAP(n_neighbors=10, n_components=2, min_dist=0.0, metric='cosine').fit_transform(embeddings)

topic_model.visualize_documents(src, embeddings=embeddings)

topic_model.visualize_documents(src, reduced_embeddings=reduced_embeddings)

topa_df = topic_model.get_document_info(src)

topa_df

topa_org_df = data.copy()
topa_org_df.reset_index(drop=True, inplace=True)
topa_df.reset_index(drop=True, inplace=True)

topa_df

tot_topa_df = pd.concat([topa_org_df, topa_df], axis = 1)

tot_topa_df

tot_topa_df[(tot_topa_df['Topic'] ==0)&(tot_topa_df['Representative_document'] ==True)]

data

sample_dropnull_df = sample_df[['q_id', 'q_creationdate', 'a_id']].groupby(['q_id', 'q_creationdate']).head(1)
sample_dropnull_df = sample_dropnull_df[['q_id', 'a_id']]

tot_topa_df = pd.merge(tot_topa_df, sample_dropnull_df, how='left', left_on = 'q_id', right_on = 'q_id')

tot_topa_df[tot_topa_df['q_id'] == 77394187]

tot_topa_df

data

tot_topa_df['date'] = pd.to_datetime(tot_topa_df['q_creationdate']).dt.date

tot_topa_df

tot_topa_not_answered_df = tot_topa_df[tot_topa_df['a_id'].isna()]
tot_topa_answered_df = tot_topa_df[tot_topa_df['a_id'].isna()==False]

tot_topa_not_answered_rate = tot_topa_not_answered_df.groupby(['date', 'Topic']).count()[['q_id']].reset_index()
tot_topa_answered_rate = tot_topa_answered_df.groupby(['date', 'Topic']).count()[['q_id']].reset_index()

tot_topa_not_answered_sum = tot_topa_not_answered_rate.groupby(['date']).sum()[['q_id']].reset_index()
tot_topa_answered_sum = tot_topa_answered_rate.groupby(['date']).sum()[['q_id']].reset_index()

tot_topa_not_answered_rate = pd.merge(tot_topa_not_answered_rate, tot_topa_not_answered_sum, on = 'date')
tot_topa_answered_rate = pd.merge(tot_topa_answered_rate, tot_topa_answered_sum, on = 'date')

tot_topa_not_answered_rate

# topic_per_rate = tot_topa_not_answered_rate.rename({'q_id_x': 'cnt', 'q_id_y':'tot_cnt'})

tot_topa_not_answered_rate['rate'] = tot_topa_not_answered_rate['q_id_x']/tot_topa_not_answered_rate['q_id_y']*100
tot_topa_answered_rate['rate'] = tot_topa_answered_rate['q_id_x']/tot_topa_answered_rate['q_id_y']*100


tot_topa_not_answered_rate

tot_topa_not_answered_rate_pivot = pd.pivot_table(tot_topa_not_answered_rate, index = 'date', columns = 'Topic',values = 'rate').reset_index()
tot_topa_answered_rate_pivot = pd.pivot_table(tot_topa_answered_rate, index = 'date', columns = 'Topic',values = 'rate').reset_index()

tot_topa_not_answered_pivot = pd.pivot_table(tot_topa_not_answered_rate, index = 'date', columns = 'Topic',values = 'q_id_x').reset_index()
tot_topa_answered_pivot = pd.pivot_table(tot_topa_answered_rate, index = 'date', columns = 'Topic',values = 'q_id_x').reset_index()

tot_topa_not_answered_rate_pivot.fillna(0, inplace=True)
tot_topa_answered_rate_pivot.fillna(0, inplace=True)

tot_topa_not_answered_pivot.fillna(0, inplace=True)
tot_topa_answered_pivot.fillna(0, inplace=True)

tot_topa_not_answered_rate_pivot

tot_topa_not_answered_rate_pivot['sdate'] = tot_topa_not_answered_rate_pivot['date'].astype('str')
tot_topa_answered_rate_pivot['sdate'] = tot_topa_answered_rate_pivot['date'].astype('str')

tot_topa_not_answered_pivot.to_csv('./chk_not_answered.csv')
tot_topa_answered_pivot.to_csv('./chk_answered.csv')

### Problem 2b: Reformat the data and generate the stacked bar plot above.


fig, ax = plt.subplots(figsize = (10, 5))

# ax.bar(topic_per_pivot['date'], topic_per_pivot[0], width = 0.5, color='blue', label = 'Party A')
ax.bar(tot_topa_answered_rate_pivot['sdate'], tot_topa_not_answered_rate_pivot[0], color='red', label = 'Party A')
ax.bar(tot_topa_answered_rate_pivot['sdate'], tot_topa_not_answered_rate_pivot[1], bottom = tot_topa_not_answered_rate_pivot[0], color = 'orange', label = 'Party B')
ax.bar(tot_topa_answered_rate_pivot['sdate'], tot_topa_not_answered_rate_pivot[2], bottom = tot_topa_not_answered_rate_pivot[1] + tot_topa_not_answered_rate_pivot[0], color = 'green', label = 'Party B')
ax.bar(tot_topa_answered_rate_pivot['sdate'], tot_topa_not_answered_rate_pivot[3], bottom = tot_topa_not_answered_rate_pivot[2] + tot_topa_not_answered_rate_pivot[1] + tot_topa_not_answered_rate_pivot[0], color = 'yellow', label = 'Party B')
ax.bar(tot_topa_answered_rate_pivot['sdate'], tot_topa_not_answered_rate_pivot[4], bottom = tot_topa_not_answered_rate_pivot[3] + tot_topa_not_answered_rate_pivot[2] + tot_topa_not_answered_rate_pivot[1] + tot_topa_not_answered_rate_pivot[0], color = 'blue', label = 'Party B')
ax.bar(tot_topa_answered_rate_pivot['sdate'], tot_topa_not_answered_rate_pivot[5], bottom = tot_topa_not_answered_rate_pivot[4] + tot_topa_not_answered_rate_pivot[3] + tot_topa_not_answered_rate_pivot[2] + tot_topa_not_answered_rate_pivot[1] + tot_topa_not_answered_rate_pivot[0], color = 'purple', label = 'Party B')


# ax.set_xlabel('Congress', fontsize = 5, labelpad = 20)
# ax.set_xlabel('Congress', fontsize = 5)
# ax.set_ylabel('Number of Seats', fontsize = 5, labelpad = 20)
# ax.set_title('Partisan Composition of Congress', fontsize = 17.5, pad = 20)

# ax.tick_params(axis = 'both', labelsize = 5)

# ax.legend(title = "Political Party",
#           title_fontsize = 15, fontsize = 5,
#           bbox_to_anchor = (1, 0.55))

plt.xticks(rotation = 45)
plt.show();

### Problem 2b: Reformat the data and generate the stacked bar plot above.


fig, ax = plt.subplots(figsize = (10, 5))

# ax.bar(topic_per_pivot['date'], topic_per_pivot[0], width = 0.5, color='blue', label = 'Party A')
ax.bar(tot_topa_answered_rate_pivot['sdate'], tot_topa_answered_rate_pivot[0], color='red', label = 'Party A')
ax.bar(tot_topa_answered_rate_pivot['sdate'], tot_topa_answered_rate_pivot[1], bottom = tot_topa_answered_rate_pivot[0], color = 'orange', label = 'Party B')
ax.bar(tot_topa_answered_rate_pivot['sdate'], tot_topa_answered_rate_pivot[2], bottom = tot_topa_answered_rate_pivot[1] + tot_topa_answered_rate_pivot[0], color = 'green', label = 'Party B')
ax.bar(tot_topa_answered_rate_pivot['sdate'], tot_topa_answered_rate_pivot[3], bottom = tot_topa_answered_rate_pivot[2] + tot_topa_answered_rate_pivot[1] + tot_topa_answered_rate_pivot[0], color = 'yellow', label = 'Party B')
ax.bar(tot_topa_answered_rate_pivot['sdate'], tot_topa_answered_rate_pivot[4], bottom = tot_topa_answered_rate_pivot[3] + tot_topa_answered_rate_pivot[2] + tot_topa_answered_rate_pivot[1] + tot_topa_answered_rate_pivot[0], color = 'blue', label = 'Party B')
ax.bar(tot_topa_answered_rate_pivot['sdate'], tot_topa_answered_rate_pivot[5], bottom = tot_topa_answered_rate_pivot[4] + tot_topa_answered_rate_pivot[3] + tot_topa_answered_rate_pivot[2] + tot_topa_answered_rate_pivot[1] + tot_topa_answered_rate_pivot[0], color = 'purple', label = 'Party B')


# ax.set_xlabel('Congress', fontsize = 5, labelpad = 20)
# ax.set_xlabel('Congress', fontsize = 5)
# ax.set_ylabel('Number of Seats', fontsize = 5, labelpad = 20)
# ax.set_title('Partisan Composition of Congress', fontsize = 17.5, pad = 20)

# ax.tick_params(axis = 'both', labelsize = 5)

# ax.legend(title = "Political Party",
#           title_fontsize = 15, fontsize = 5,
#           bbox_to_anchor = (1, 0.55))

plt.xticks(rotation = 45)
plt.show();

idx = np.arange(tot_topa_answered_rate_pivot.shape[0])
w = 0.15

plt.figure(figsize = (10, 5))
plt.title('pctchg in topic 1')
plt.xlabel('month')
plt.ylabel('pct')
plt.plot(tot_topa_answered_rate_pivot[0], label = 'answered')
plt.plot(tot_topa_not_answered_rate_pivot[0], label = 'not answered')
plt.xticks(idx, tot_topa_answered_rate_pivot['sdate'], rotation = 30)
plt.legend(ncol = 1)
plt.show()

idx = np.arange(tot_topa_answered_rate_pivot.shape[0])
w = 0.15

plt.figure(figsize = (10, 5))
plt.title('pctchg in topic 2')
plt.xlabel('month')
plt.ylabel('pct')
plt.plot(tot_topa_answered_rate_pivot[1], label = 'answered')
plt.plot(tot_topa_not_answered_rate_pivot[1], label = 'not answered')
plt.xticks(idx, tot_topa_answered_rate_pivot['sdate'], rotation = 30)
plt.legend(ncol = 1)
plt.show()

idx = np.arange(tot_topa_answered_rate_pivot.shape[0])
w = 0.15

plt.figure(figsize = (10, 5))
plt.title('pctchg in topic 3')
plt.xlabel('month')
plt.ylabel('pct')
plt.plot(tot_topa_answered_rate_pivot[2], label = 'answered')
plt.plot(tot_topa_not_answered_rate_pivot[2], label = 'not answered')
plt.xticks(idx, tot_topa_answered_rate_pivot['sdate'], rotation = 30)
plt.legend(ncol = 1)
plt.show()

idx = np.arange(tot_topa_answered_rate_pivot.shape[0])
w = 0.15

plt.figure(figsize = (10, 5))
plt.title('pctchg in topic 4')
plt.xlabel('month')
plt.ylabel('pct')
plt.plot(tot_topa_answered_rate_pivot[3], label = 'answered')
plt.plot(tot_topa_not_answered_rate_pivot[3], label = 'not answered')
plt.xticks(idx, tot_topa_answered_rate_pivot['sdate'], rotation = 30)
plt.legend(ncol = 1)
plt.show()

idx = np.arange(tot_topa_answered_rate_pivot.shape[0])
w = 0.15

plt.figure(figsize = (10, 5))
plt.title('pctchg in topic 5')
plt.xlabel('month')
plt.ylabel('pct')
plt.plot(tot_topa_answered_rate_pivot[4], label = 'answered')
plt.plot(tot_topa_not_answered_rate_pivot[4], label = 'not answered')
plt.xticks(idx, tot_topa_answered_rate_pivot['sdate'], rotation = 30)
plt.legend(ncol = 1)
plt.show()

tot_topa_answered_rate_pivot

tot_topa_answered_rate_pivot
